fix: moved initData in  mockedData.ts file
