+++ src/components/LineChart/LineChart.stories.mdx
@@ -1,7 +1,8 @@
 import { Meta, Preview, Props, Story } from '@storybook/addon-docs/blocks';
 import { withKnobs, boolean, array, select, text } from "@storybook/addon-knobs";
 import LineChart from './LineChart';
-import { color, initData } from './story.utils'
+import { color } from './story.utils';
+import { initData } from './mockedData';
 
 <Meta title="Design System/Chart" component={LineChart} />
